# frontend/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AhoraMismo": {
        "properties": {
          "alertasDeFraude": {
            "format": "int64",
            "type": "integer"
          },
          "comandasAbiertas": {
            "format": "int64",
            "type": "integer"
          },
          "eventosOutboxEnError": {
            "format": "int64",
            "type": "integer"
          },
          "eventosOutboxPendientes": {
            "format": "int64",
            "type": "integer"
          },
          "insumosBajoMinimo": {
            "format": "int64",
            "type": "integer"
          },
          "mesasActivas": {
            "format": "int64",
            "type": "integer"
          },
          "mesasOcupadas": {
            "format": "int64",
            "type": "integer"
          },
          "pagosPorAcreditar": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "MesaResponseDto": {
        "properties": {
          "activa": {
            "type": "boolean"
          },
          "capacidad": {
            "format": "int32",
            "type": "integer"
          },
          "estado": {
            "enum": [
              "LIBRE",
              "OCUPADA",
              "RESERVADA",
              "INHABILITADA"
            ],
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "numero": {
            "type": "string"
          },
          "reservadaANombreDe": {
            "type": "string"
          },
          "reservadaPara": {
            "format": "date-time",
            "type": "string"
          },
          "zona": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Operacion": {
        "properties": {
          "comandasCanceladas": {
            "format": "int64",
            "type": "integer"
          },
          "comandasFraudulentas": {
            "format": "int64",
            "type": "integer"
          },
          "minutosObjetivoCocina": {
            "format": "int32",
            "type": "integer"
          },
          "minutosPromedioCocina": {
            "format": "double",
            "type": "number"
          }
        },
        "type": "object"
      },
...
      "PorCanal": {
        "properties": {
          "canal": {
            "enum": [
              "POS",
              "WHATSAPP_BOT",
              "DISCORD_BOT",
              "WEB"
            ],
            "type": "string"
          },
          "cantidadComandas": {
            "format": "int64",
            "type": "integer"
          },
          "total": {
            "type": "number"
          }
        },
        "type": "object"
      },
      "PorEstado": {
        "properties": {
          "comandas": {
            "format": "int64",
            "type": "integer"
          },
          "estado": {
            "enum": [
              "ENCOLADO",
              "EN_PREPARACION",
              "EN_DESPACHO",
              "ENTREGADO",
              "PAGADO",
              "CONCLUIDO",
              "CANCELADO",
              "FRAUDULENTO"
            ],
            "type": "string"
          },
          "total": {
            "type": "number"
          }
        },
        "type": "object"
      },
...
      "ProductoTopDto": {
        "properties": {
          "montoTotal": {
            "type": "number"
          },
          "platillo": {
            "type": "string"
          },
          "unidadesVendidas": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "Punto": {
        "properties": {
          "comandas": {
            "format": "int64",
            "type": "integer"
          },
          "inicio": {
            "format": "date-time",
            "type": "string"
          },
          "total": {
            "type": "number"
          }
        },
        "type": "object"
      },
...
      "ReporteVentasDto": {
        "properties": {
          "desde": {
            "format": "date-time",
            "type": "string"
          },
          "hasta": {
            "format": "date-time",
            "type": "string"
          },
          "porCanal": {
            "items": {
              "$ref": "#/components/schemas/PorCanal"
            },
            "type": "array"
          },
          "totalVendido": {
            "type": "number"
          }
        },
        "type": "object"
      },
...
      "SerieVentasDto": {
        "properties": {
          "desde": {
            "format": "date-time",
            "type": "string"
          },
          "granularidad": {
            "enum": [
              "HORA",
              "DIA"
            ],
            "type": "string"
          },
          "hasta": {
            "format": "date-time",
            "type": "string"
          },
          "puntos": {
            "items": {
              "$ref": "#/components/schemas/Punto"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "TableroDto": {
        "properties": {
          "ahoraMismo": {
            "$ref": "#/components/schemas/AhoraMismo"
          },
          "desde": {
            "format": "date-time",
            "type": "string"
          },
          "hasta": {
            "format": "date-time",
            "type": "string"
          },
          "operacion": {
            "$ref": "#/components/schemas/Operacion"
          },
          "porEstado": {
            "items": {
              "$ref": "#/components/schemas/PorEstado"
            },
            "type": "array"
          },
          "ventas": {
            "$ref": "#/components/schemas/Ventas"
          }
        },
        "type": "object"
      },
...
      "Ventas": {
        "properties": {
          "comandas": {
            "format": "int64",
            "type": "integer"
          },
          "ticketPromedio": {
            "type": "number"
          },
          "total": {
            "type": "number"
          }
        },
        "type": "object"
      },
      "VentasPorMozoDto": {
        "properties": {
          "comandas": {
            "format": "int64",
            "type": "integer"
          },
          "mozo": {
            "type": "string"
          },
          "mozoId": {
            "format": "uuid",
            "type": "string"
          },
          "ticketPromedio": {
            "type": "number"
          },
          "total": {
            "type": "number"
          }
        },
        "type": "object"
      },
...
      "VinculacionBotDto": {
        "properties": {
          "activo": {
            "type": "boolean"
          },
          "cargoNombre": {
            "type": "string"
          },
          "correo": {
            "type": "string"
          },
          "discordUserId": {
            "type": "string"
          },
          "nombreCompleto": {
            "type": "string"
          },
          "trabajadorId": {
            "format": "uuid",
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API de POS, inventario, comandas, cocina, despacho y caja del sistema de gestion gastronomica Chaquena.",
    "title": "Backend Logistica - Chaquena",
    "version": "v1"
  },
  "openapi": "3.0.1",
  "paths": {
...
    "/api/v1/bots/vinculaciones/{trabajadorId}": {
      "delete": {
        "description": "El trabajador sigue de alta y sigue entrando al front; solo pierde el acceso por chat. Es lo que hay que hacer cuando alguien se vinculo con la cuenta equivocada o deja el local.",
        "operationId": "desvincularCuentaBot",
        "parameters": [
          {
            "in": "path",
            "name": "trabajadorId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VinculacionBotDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Liberar la cuenta de mensajeria de un trabajador",
        "tags": [
          "Bots"
        ]
      }
    },
...
    "/api/v1/mesas/{id}/liberar": {
      "post": {
        "operationId": "liberar",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MesaResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Liberar la mesa",
        "tags": [
          "Salon - Mesas"
        ]
      }
    },
...
    "/api/v1/reportes/productos-top": {
      "get": {
        "operationId": "productosTop",
        "parameters": [
          {
            "in": "query",
            "name": "desde",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hasta",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limite",
            "required": false,
            "schema": {
              "default": 10,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProductoTopDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Platillos mas vendidos, para decidir la carta",
        "tags": [
          "Reportes"
        ]
      }
    },
...
    "/api/v1/reportes/serie-ventas": {
      "get": {
        "operationId": "serieVentas",
        "parameters": [
          {
            "in": "query",
            "name": "desde",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hasta",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "granularidad",
            "required": false,
            "schema": {
              "default": "DIA",
              "enum": [
                "HORA",
                "DIA"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SerieVentasDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Ventas repartidas por hora o por dia, con los bloques vacios incluidos",
        "tags": [
          "Reportes"
        ]
      }
    },
    "/api/v1/reportes/tablero": {
      "get": {
        "description": "Sin rango responde por el dia en curso. Las cifras del bloque 'ahoraMismo' son una foto del presente y no dependen del rango pedido.",
        "operationId": "tableroLocal",
        "parameters": [
          {
            "in": "query",
            "name": "desde",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hasta",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableroDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "El local de un vistazo: venta, cocina, salon, caja y eventos",
        "tags": [
          "Reportes"
        ]
      }
    },
    "/api/v1/reportes/ventas": {
      "get": {
        "operationId": "ventas",
        "parameters": [
          {
            "in": "query",
            "name": "desde",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hasta",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReporteVentasDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Ventas por rango de fechas y canal de origen",
        "tags": [
          "Reportes"
        ]
      }
    },
    "/api/v1/reportes/ventas-por-mozo": {
      "get": {
        "operationId": "ventasPorMozo",
        "parameters": [
          {
            "in": "query",
            "name": "desde",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hasta",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VentasPorMozoDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Cuanto vendio cada mozo en el rango",
        "tags": [
          "Reportes"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}